Add more theme adocs and explicitly use Path Parameter when needed
The theme tests had to be redesign to prevent duplication of document generation.
The use of `optional()` and `ignored()` is necessary in some cases.
When `requestFields()` or `responseFields()` are specified, all fields must be document or explicitly set to ignored.
This may not be the case with `RequestParameters()`.

Request Parameters and Requests Fields should both be called "Request Parameter", but "Path Parameter" needs its own label.