73: Update selectedPublication Resovler to use syncIds
Test concept of finding all related documents at once via existing
field. This should allow us to add in pagination, filtering and faceting
if need be.

Move retrieval out of constructor so it is called only when queried.

To be accurate new multivalue id fields would need to be added to the
Document index for each type of relationship: authorIds, editorIds, etc.