[136] Add complete author list to person.
This does:
- Add solr schema field type `ordered_whole_strings`, based off of `nested_whole_strings`.
- The `ordered_whole_strings` should support nested strings within it but the java implementation currently does not handle this.
- Add `ORDERED_DELIMITER` discovery constant.
- Add boolean annotation property `ordered` to `PropertySource`.
- Add `OrderedComparator` class under a new comparator directory within the discovery directory tree.
- Add `completeAuthorList` to `Document` as an `ordered_whole_strings` array, utilizing the newly added `ordered` `PropertySource`.
- Provide basic implementation for ordering the list and removing the ordering digits.
- Add `completeAuthorList` sparql file.

This takes a simple approach so that future iterations will have an easier time updating this.
The provided sparql regex query is guaranteeing a certain consistency in the output.
This does not currently validate the structure on the java side but I recommend doing this eventually.
The ordered property source currently just continues, but this should be eventually improved to handle nested strings.